# contracts/java-python-rag/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Sage Vault Java-Python RAG Contract
  version: 1.0.0
paths:
  /internal/v1/answers:
    post:
      operationId: answer
      parameters:
        - {name: X-Sage-Timestamp, in: header, required: true, schema: {type: string}}
        - {name: X-Sage-Signature, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
      responses:
        '200':
          description: Project-owned SSE events
          content:
            text/event-stream:
              schema: {type: string}
        '401': {description: Invalid deployment signature}
  /internal/v1/indexing:
    post:
      operationId: indexDocument
      parameters:
        - {name: X-Sage-Timestamp, in: header, required: true, schema: {type: string}}
        - {name: X-Sage-Signature, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexingCommand'
      responses:
        '202':
          description: Command accepted; processing happens asynchronously
        '401': {description: Invalid deployment signature}
  /internal/v1/indexing/callbacks:
    post:
      operationId: indexingCallback
      parameters:
        - {name: X-Sage-Timestamp, in: header, required: true, schema: {type: string}}
        - {name: X-Sage-Signature, in: header, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexingCallback'
      responses:
        '200':
          description: Callback accepted and idempotently reconciled
        '401': {description: Invalid deployment signature}
components:
  schemas:
    AnswerRequest:
      type: object
      additionalProperties: false
      required: [knowledgeBaseId, question, requestId, generationId]
      properties:
        knowledgeBaseId: {type: integer, format: int64, minimum: 1}
        question: {type: string, minLength: 1}
        requestId: {type: string, minLength: 1}
        generationId: {type: string, format: uuid}
    IndexingCommand:
      type: object
      additionalProperties: false
      required: [taskId, attempt, knowledgeBaseId, documentId, filename, sourceUrl, requestId]
      properties:
        taskId: {type: string, minLength: 1}
        attempt: {type: integer, minimum: 1}
        knowledgeBaseId: {type: integer, format: int64, minimum: 1}
        documentId: {type: string, minLength: 1}
        filename: {type: string, minLength: 1}
        sourceUrl: {type: string, format: uri, minLength: 1}
        requestId: {type: string, minLength: 1}
    IndexingCallback:
      type: object
      additionalProperties: false
      required: [taskId, attempt, documentId, success, chunksCount, requestId]
      properties:
        taskId: {type: string, minLength: 1}
        attempt: {type: integer, minimum: 1}
        documentId: {type: string, minLength: 1}
        success: {type: boolean}
        chunksCount: {type: integer, minimum: 0}
        requestId: {type: string, minLength: 1}
        diagnostics:
          type: object
          additionalProperties: true